+++ client/src/components/Header/styled.ts
@@ -34,23 +34,3 @@ export const Nav = styled.nav`
   justify-content: space-between;
   align-items: center;
 `;
-
-export const NavLeftList = styled.ul`
-  padding-left: 3.125rem;
-  gap: 2.4rem;
-  display: flex;
-  align-items: center;
-  font-weight: 600;
-`;
-
-export const NavRightList = styled.ul`
-  padding-right: 3.125rem;
-  gap: 1rem;
-  display: flex;
-  align-items: center;
-  font-weight: 800;
-`;
-
-export const Login = styled.li`
-  padding: 0 1.4rem;
-`;
