+++ client/src/components/NavLeftList/index.ts
@@ -1,0 +1,1 @@
+export * from "@/components/NavLeftList/NavLeftList";
