+++ client/src/components/NavLeftList/styled.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const NavList = styled.ul`
+  padding-left: 3.125rem;
+  gap: 2.4rem;
+  display: flex;
+  align-items: center;
+  font-weight: 600;
+`;
