+++ client/src/components/NavRightList/styled.ts
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const NavList = styled.ul`
+  padding-right: 3.125rem;
+  gap: 1rem;
+  display: flex;
+  align-items: center;
+  font-weight: 800;
+`;
+
+export const Login = styled.li`
+  padding: 0 1.4rem;
+`;
